+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [6.2.1] - 2022-02-04
+### Fixed
+- Ensures that the UniswapPoolScanner agent can start without OTA enabled
+- Uniswap Verified Pool optimized query index
+
 ## [6.2.0] - 2022-01-28
 ### Added
 - Uniswap OTA updates